ipal version 2014100100
This version is for Moodle 2.7.1. It will not run with Moodle 2.6. It
has cleaner code as well as being suitable for Moodle 2.7.1.  It
supports in-class polling when students use web-enabled devices and/or
smart phones with free Apps. There is a polling App for the iPhone and
iPad (the IOS version) and an App for Android devices.