Create README.md
This version of IPAL (2014100100) is compatible with Moodle 2.7.1. 
It will not run in Moodle 2.6.
The directory ipal should be placed in the moodle mod folder.
It supports in-class polling when students use web-enabled devices and/or smart phones with free Apps. 
There is a polling App for the iPhone and iPad (the IOS version) and an App for Android devices.