Improve ipal attendance utility
The utility that allows ipal to create attendance questions and then
update the record in the attendance module is improved. There is one
less separate web page that the teacher has to use. Mobile access does
not need to be turned on.. Some of the text has been moved to the
language file for easier translation.