Ignore multiple records in findinstructor()
The call to $DB->get_record_sql() in findinstructor() was generating an
error when there is more than one editing teacher in the course.  This is
a quick workaround - setting the IGNORE_MULTIPLE flag so that the first
matching record is returned with no error.  It may be desirable to rework
this function to pick from multiple teachers in a smarter way.