Allow non-alphanumeric characters in user name for IPAL Apps
A change (PARAM_ALPHANUMEXT to  PARAM_RAW_TRIMMED in line 39) was made
to the tempview.php file so that user names could include
non-alphanumeric characters. Some schools have started using email
addresses (containing the . and @ characters) for user names. This
breaks the login using the IPAL Apps. This change fixes that.