Fix bug in mobile polling
A bug was introduced when removing MSWord formatting from questions. This bug affected the recording of student responses with mobile devices. The variable $question was changed to $myquestion in lines 928-931 in ipal/locallib.php. This bug was not present in any version published on the moodle.org site and has been removed from future versions.