Add nocurrentquestion for App
When there is no current question (polling has stopped and/or no
question has been sent) the Add needs to receive the <p tag with the
id='questiontype' and the text of nocurrentquestion so that the App
displays this message to the user. This tag was added to locallib.php in
line 968.