view.php changed to check for using quizzes
If IPAL uses a quiz, the ipal->mobile field = 4. view.php now checks this value and, if this field is = 4, checks to see if a quiz is being selected (a submitted value of $quizid). If so, a label is created for Mobile Moodle users.. If this provides a quizid or if a quizid is already provided, the teacher is sent to the /ipal/liveview/quizview.php?n=$ipal->quizid page. If a student goes to this page, if the IPAL instance is ready to run, the student is sent to the /ipal/liveview/quiz_iframe.php?quizid=$ipal->quizid page.